refactor(view): move InputView to view package to align UI components
